Minimal type reconstruction
Reconstruction of basic types for TLA+ proof obligations. This pass over
the PO adds sort annotations to bound variables and insert coercions
where needed. The resulting PO is suitable for a direct, faithful
encoding into typed format such as SMT-LIB.

The sorts are among set, bool, int, real and string. Types range over
all possible types for:
- constants (sort 's');
- first-order operators ('s X .. X s -> s');
- second-order operators ('t X .. X t -> s' where 't' is a FO type).

Everything can be coerced to 'set', everything can be coerced to 'bool'.

In cases where expressions of different sorts must be unified, they are
all coerced to 'set'. For example:
    IF 0 = 1 THEN TRUE ELSE {}
is transformed into:
    IF 0 = 1 THEN BoolToSet(TRUE) ELSE {}
(The other possibility was to coerce '{}' to a boolean expression.)

Type annotations are bound (as properties) to bound variables in
quantifiers (\A, \E, CHOOSE, LAMBDA, etc.) and declarations (operator
definitions, etc.)