Sorts and functions collection
Declaration of sorts, functions and axioms are handled in
'backend/smttable'. This might change in the future; it might be simpler
to apply local transformations to the sequent (while adding axioms) in the
'reduce' phase of the encoding, and then let 'backend/smtlib' implement
a straightforwad translation, with no data collection.

Most axioms are missing for now.