Disambiguation: opaques replaced by annotations
In particular, different versions of '+', '-' etc. are still represented
with the builtin Plus constructor; they are distinguished by an
annotation.  The int and real versions have an annotation; no annotation
means the set version.

Opaques are still inserted for casts and anys.  These opaques get a
special annotation; they should be recognized using this annotation
rather than with a string comparison.