Free variables function fixed
Assume a local context with only a variable x in it.  The previous
implem called on '\A y : x' would have returned the free variable x, but
as the index 2 and not 1.  This is because 2 is the actual index in the
raw expression.  But 1 is the index that makes sense relative to the
context of the call.

This has been corrected, but now it no longer makes sense to implement a
function that returns the bound variables as De Bruijn indexes.
(Actually it did not make sense before either.)