Relocalize expressions for SetSt axiom
This is very complicated.  The idea is that we have an expression like:
{ x \in S : P(x, y, z, a, b, c) } where the variables x, y, z are local
(quantifiers) and the variables a, b, c are global (declarations).  The
construction must be replaced by a first-order application F(S, y, z)
and lead to the new axiom:
    \A s, a1, a2, x : x \in F(s, a1, a2) iff P(x, a1, a2, a, b, c)

But 1) the axiom cannot be inserted above any of the declarations of a,
b or c, 2) the new local context [s, a1, a2, x] is very different from
the original local context [y, z, x].  Solving 2) involves cumbersome
renamings (implemented in the function "badaboom").  Solving 1) is less
complicated.

I'm afraid there is no self-explanatory name for the function "badaboom"