Changed symbol conventions
New datatype tla_smb to represent standard symbols in a more organized
way.  The old smb type is still used, but for a more generic
representation.  An optional tla_smb can be attached to a smb as a
definition.  Encode.Canon still attaches symbols of the smb datatype,
but it should be fixed later to also attach a definition.