Encode.Table: new architecture
Type for symbols is monomorphic: it's impossible to annotate exprs with
polymorphic data.  Encode.Table is now purely abstract; the data on
symbols is put separately in Encode.Data, not visible outside of
encode/.  Symbols have a definition attached, and some optional info on
a typed version for the symbol.