BUG: keep names with nesting depth
The renaming of bound identifiers during lambdification
does not take into account the nesting depth of `ENABLED` and `\cdot`.
These names unintentionally propagated to replacement by quantifiers
(the proof directive `ExpandENABLED`), when replacement by quantifiers
was changed to two phases (lambdification followed by replacement with
quantifiers).

This change reverts to using the identifiers that are indexed by
nesting depth when replacing `ENABLED` and `\cdot` by quantifiers.
The naive naming of bound identifiers is now used in the lambdified
form only when the proof directive `Lambdify` is used.

This change follows a comment in pull request:
https://github.com/tlaplus/tlapm/pull/9