MAI: rm TODO
The context in the bounds of rigid quantification,
set comprehension, and function constructors is
the context before each construct. So bumping is
not needed, because the newly introduced symbols
cannot be used in the other bounds. This conclusion
has been confirmed using SANY.

This change follows a comment in pull request:
https://github.com/tlaplus/tlapm/pull/9