BUG: correct fingerprinting of variables
Variables were fingerprinted as `VAR(i)` where `i` an index,
and bound identifiers also with `VAR(j)` where `j` an index
produced using a different counter. The counter for variables
is `counthyp`, and the counter for bound constants and bound
variables is `countvar`.

When the two counters are both equal to 1,
`VAR(1)` results for both the occurrence of a variable, and
the occurrence of a constant bound by a rigid quantifier.

This commit adds a test file that catches this error.