BUG: fingerprint `AutoUSE` results
This change makes fingerprints take into account
the result of the proof directive AutoUSE.

Previously, fingerprinting was done before any
expansion of definitions, normalization, and
automated expansion of definitions, expansion of
ENABLED and of \cdot.

This approach worked correctly for BY DEF definitions,
because those definitions were included in the
fingerprint as context.

With this change, only for proof obligations that
include the proof directive AutoUSE, the fingerprint
is computed after expansion of definitions, normalization,
automated expansion of definitions, expansion of ENABLED,
and of \cdot. This change ensures that the proof
obligation is fingerprinted with all automatic
expansions of definitions applied.

A test is added that catches this error.