BUG: correctly shift module units when instantiating
When instantiating inside a LET, definitions are kept,
and other module units omitted. A negative shift is
applied to the remaining module units. This negative shift
should equal minus the number of hypotheses that the
omitted module unit introduces in the context.

The number of hypotheses for each module unit is
computed by the function `M_t.hyps_of_modunit`.