API: update to keywords of Isabelle 2021
The script `isabelle_keywords_get.sh` is outdated,
because the file it gets the keywords from is not
part of recent Isabelle distributions.

Instead, a Python script `isabelle_keywords_update.py`
has been added, which relies on the Scala script
`outer_keywords.scala`, which gets the Isabelle 2021
keywords for the logic `Pure`.