BUG: correctly interpret multiple `EXCEPT` assignments
`List.map` was used to iterate over the assignments that
appear in an `EXCEPT` constructor. However, such assignments
mean nesting of `EXCEPT` constructors.

This change corrects this behavior, by replacing `List.map`
with a recursion implemented with `List.fold_left` to nest
the `EXCEPT` constructors, and using as the function `f` the
latest result of this nesting, which gets substituted when
forming the `@` expressions using the same code as before.