MAI: raise custom exception within function `reduce_one`
to avoid the OCaml compiler warning (with version 4.12.0):

```
&& ocamlopt -annot -g -warn-error +1+2+5+6+8+10..26+29..31+36  -I . -I backend -I frontend -I expr -I module -I proof -I util -I pars -I typesystem -I smt -I ../backend -I ../frontend -I ../expr -I ../module -I ../proof -I ../util -I ../pars -I ../typesystem -I ../smt -I .. -c `basename pars/pco.ml`
File "pco.ml", line 736, characters 24-36:
736 |               | Failure "reduce_one" -> fail "incomplete expression"
                              ^^^^^^^^^^^^
Warning 52 [fragile-literal-pattern]: Code should not depend on the actual values of
this constructor's arguments. They are only for information
and may change in future versions. (See manual section 9.5)
```

This warning was raised because `failwith` raises a `Failure`,
and the OCaml documentation warns to not pattern match on the
string literal because it may change in future versions.