TST: `INSTANCE TLAPS`
to ensure that no warnings or errors are raised when
instantiating the module `TLAPS`. This test is motivated
by the warnings about `ENABLED` applied to `ACTION` operators
that would be raised when applying the current version of
`tlapm` to versions of the TLA+ module `TLAPS` from older
TLAPS releases.

Also, add to the TLAPS website page "What's new" an entry
for the next release of TLAPS (version TBD) that notes the
possibility of this warning, and what to do in that case.

https://github.com/tlaplus/tlapm/issues/15