REF: use lowercase for website paths
This commit relocates several binary files, but does not
change their contents. So `git` is not going to store any
new blob objects for the contents of these binaries.

The only change by `git` is that it will create new tree
objects, to represent the rearranged directories (and
a new commit object to represent this commit). These tree
objects are lightweight.

In addition, these new tree objects would have been created
anyway due to relocating other files in the same directory
as the binary files, even if the binary files were not
relocated, due to the change in that directory's contents.

Relevant documentation:
https://git-scm.com/book/en/v2/Git-Internals-Git-Objects

Also, this commits converts the website menu names to
lowercase, in order to obtain paths, in the script
`doc/web/assets/header.js`.

For the path changes in this commit to take effect on macOS,
please either run:

```shell
git config --global core.ignorecase false
```

and then checkout a commit before this commit, and then
checkout this commit, or clone the `tlapm` repository anew.