MNT: uniformize HTML headers of content webpages
"content webpages" here refers to pages that do not result in
redirection. The HTML headers of redirection pages differ with
each other in what the target of redirection is, so those
headers will remain in place, and cannot be made identical.

Headers of all other webpages, and also all the footers,
will be removed from the Markdown files, and added dynamically
before calling `pandoc` to generate the HTML files.

This commit removes lines that caused differences in HTML
headers among files. After these changes, all Markdown files
for webpages that are not redirections have identical headers,
and identical footers. All Markdown files of webpages that are
redirections have identical footers. These checks have been
performed with a Python script.

The lines removed by this commit are going to be added back
to the HTML header, in the upcoming commit that will remove
the HTML header from all Markdown files that are sources of
content pages. In this way, these lines will overall be
added to all content webpages.

The reason these lines were removed here, instead of adding
them to all Markdown files, is to reduce the amount of
files changed (and blobs created by `git`), since those
headers are anyway going to be removed, and generated
automatically.