BUG: correct comment that appears to be in error
The error appears to be that the function
`Smt.Ectx.to_fresh` creates `Fresh` values
using the constructor `Shape_expr`,
not the constructor `Shape_op`.

The correction refers to the function `Smt.Ectx.hack_bs`,
instead of the function `to_fresh`, based on a matching
comment that is located on the line above the function
`hack_bs`.