BUG: do not match `Fact` as if it was named "_"
because doing so could cause a De Bruijn index being
created for a fact. This change is to the functions
`Expr.Anon.hyp_is_named` and `Expr.T.find_hyp_named`.

The function `Expr.T.hyp_name` is changed accordingly,
in order to make it less likely that a `Fact` name
is accidentally matched in code that uses `hyp_name`.

The function `hyp_name` returns a string, and is called
from the function `find_hyp_named` also for `Fact`s,
therefore an `assert false` cannot be placed in the
pattern case for `Fact` inside the function `hyp_name`.

Also, the string returned for facts from the function
`hyp_name` is expected to help in debugging, and when
printing information.

I am not aware of the underscore being used in TLA+
syntax to refer to facts.
Underscores do not appear to be used in this way
in any TLA+ file in the `git` repository of `tlapm`.
Underscores do not appear to be used in this way
in any OCaml files of `tlapm`.