REF: replace `\r\n` with `\n`
In preparation for adding a `.gitattributes` file
in an upcoming commit. Note that the line endings
need to be converted before the `.gitattributes`
file appears.

Otherwise, `git` enters an unstable state upon
checkout, where files that have not changed
appear to have changed, due to how filtering
of line endings is done by `git`.

In more detail, when any files are committed
with `\r\n` line endings, and a `.gitattributes`
is present, then upon checkout of that commit,
`git` will create a file with `\r\n` line endings.

When `git` compares the checked-out file with
the committed one, `git` converts the `\r\n`
line endings to `\n`, due to the presence of
the `.gitattributes` file.

So `git` concludes that the files changed,
even though that is a consequence of
checking out without converting line endings,
and converting the line endings when looking
for changed files.

This situation can arise only at commits
that include both a `.gitattributes` file,
and a file with `\r\n` line endings.

This situation does not arise before
a `.gitattributes` file is committed, or
after all files have been converted to
have `\r\n` line endings.
Thus, this situation is avoided by
converting line endings before adding
a `.gitattributes` file.

Without this sequencing, any checkout
of an intermediate commit would cause
uncommitted changes to appear, and
require `git checkout -f ...` to
transition to another commit.

<https://git-scm.com/docs/gitattributes>