BUG: compare proof's filename with adjusted `tlapm` args
For each proof, `tlapm` was deciding whether to generate
proof obligations by comparing the filename stored in the
proof's location with each of the filenames given to
`tlapm` as command-line arguments (this comparison happens
inside the function `Proof.Gen.generate`).

This was an error, because `tlapm` appends the string
".tla" to those filenames (more accurately the paths)
among the command-line arguments that do not end with
".tla". So the proof's filename would end in ".tla"
when the command-line argument did not end in ".tla".

In that case, `tlapm` skipped generating proof obligations,
and exited as if all proof obligations had been proved.
This behavior resulted when the following condition was true:

```
/\ "filename" is given as an argument to `tlapm`
/\ "filename" does not end with ".tla"
/\ a file with name equal to the concatenation of
   "filename" and ".tla" does exist
```

This commit changes module `Tlapm` to apply the mapping
that appends ".tla" to the command-line arguments that
do not end with ".tla", in order for the filenames
(more accurately: file basenames) in the
list `!Params.input_files` to correspond to the filenames
of the files that `tlapm` actually searches for in the
filesystem.

With this change, the comparison inside the function
`Proof.Gen.generate` is comparing the filename stored
in the proof's location with the filenames that `tlapm`
looked for, which may differ from the positional
command-line arguments given to `tlapm`, depending on
the ending of each command-line argument.

This commit adds to function `Proof.Gen.generate`
relevant informational messages that are printed
only when the nonpositional command-line parameter
`--verbose` is given to `tlapm`.

This commit does not change the function
`Proof.Gen.generate` in any way other than
printing messages.

A motivation for including the change to
the module `Proof.Gen` in the same commit
with the change to the module `Tlapm` is
that they are related, and also as a means
to indicate where the comparison of filenames
occurs inside `Proof.Gen`.