BUG: rewrite method `Proof.Visit.map#usable`
and move the previous implementation of the method
`Proof.Visit.map#usable` to the class `Proof.Anon.anon`,
namely as the method `Proof.Anon.anon#usable`,
which overrides there the newly rewritten method
`Proof.Visit.map#usable`.

In other words, this commit adds the method
`Proof.Anon.anon#usable`, which is
code equivalent to the previous implementation of
the method `Proof.Visit.map#usable`.
Therefore, the class `Proof.Anon.anon` remains
unchanged.

This commit does change the method `Proof.Visit.map#usable`,
for the pattern-case of `dw.core` matching `Dvar v` (where `v`
is a `string`). This is the case of the name of a defined
operator listed within a `BY DEF`. This pattern-case arises
only after conversion to De Bruijn indices.

It appears that `Proof.Visit.map` had been used only in
`Proof.Anon.anon`, and for syntax trees that have already
been converted to De Bruijn indices. This is why this
error did not result in the printing of any warnings
in the past (from the `Errors.warn`).

It was impossible for the previous implementation of
the method `Proof.Visit.map#usable` to have worked for
the case of a string identifier (instead of an integer
index) inside the given definitions.

The reason is that any such string was used to construct
the expression `Opaque v` (wrapped), and apply method
`self#expr` to this expression. The result was expected
to be a De Bruijn index. This is impossible to happen,
because the class `Proof.Visit.map` subclasses the class
`Expr.Visit.map`, and the method `Expr.Visit.map#expr`
maps the expression `Opaque v` to `Opaque v` (wrapped).
No conversion to De Bruijn indices occurs inside the
method `Expr.Visit.map#expr`.