CI: test using `pr.yml` when file `pr.yml` changes
Running the test configuration that is defined in
the file `.github/workflows/pr.yml` whenever
the file `pr.yml` changes enables testing changes to
the configuration file `pr.yml`, during development of
the file itself.

This change is motivated by a typo when changing the
file `pr.yml`, which caused the file `pr.yml` to have
a semantic error (an identifier occurred that was
defined in `env`, but was referenced without the
prefix `env.`).

This typo was not detected during pushes to branches,
and caused an error in CI when a pull request was
opened, because GitHub Actions then attempted to
run `pr.yml` for the first time.

Checking with `yamllint` did not catch this error,
because this error requires interpretation by
GitHub Actions.

Two ways to test `pr.yml` during development are:

- to open pull requests in another repository,
  only for the purpose of testing

- to have local runners of GitHub Actions,
  and simulate a pull-request event there.

These are more accurate ways of testing.
Testing the file `pr.yml` in a branch (as done
by this commit) does not test for the
case of opening a pull request.

Currently, the file `pr.yml` runs the same
activities when triggered by a `push` event,
as when triggered by the opening of a pull request.
So testing in a branch is simple and
may suffice for now.


NOTE:
- each event name is followed by a colon (`:`)
  <https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#using-activity-types-and-filters-with-multiple-events>
- paths of changed files are specified for `push` events
  - <https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#onpushpull_requestpull_request_targetpathspaths-ignore>
  - <https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#using-filters>