CI: `cat` logs (`*.err` and `*.out`) of test files
upon failure of any tests:
- in the fast collection of tests, or
- in the complete collection of tests

Thus, the logs shown will correspond to
the collection of tests that failed.

The newly added CI step that displays the
log files is located after all steps that
run tests (so both the fast tests, and
the complete collection of tests).

The reason for placing this logging CI step
at the end, is so that the logging step
will `cat` failed tests in all cases that
a preceding CI step (one that runs tests) failed.