CI: always `cat` the file `test/tests.log`
This change shows `tests.log` also when the tests fail.

The file `tests.log` is expected to be present when
either the fast collection of tests, or the complete
collection of tests, has failed.

This is because in both cases, the tests are run by
the script `test/TOOLS/do_tests`, which defines:

```sh
export LOGFILE="$BASEDIR/tests.log"
```

[Documentation relevant to the expression
`if: always()` for GitHub Actions](
    https://docs.github.com/en/actions/learn-github-actions/expressions#always)